feat(models): move 8 per-model policy subclasses to tolokaforge-models wheel with deprecation shim (Milestone 29 Stage 2)
Move eight per-model policy subclasses from the engine's LLM layer to
``tolokaforge_models.policies.<family>``:

* ``GeminiSchema``, ``GeminiRecursiveSchema``, ``ScalarArrayDictMapResponse``
  → ``tolokaforge_models.policies.gemini``
* ``JsonRecursiveCoerceResponse``, ``ItemRecursiveUnwrapResponse``,
  ``MinimaxM3TagRecoveryResponse`` → ``tolokaforge_models.policies.minimax``
* ``RefResolvingDictMapHints`` → ``tolokaforge_models.policies.inkling``
* ``OpenAISummaryReplayReasoningCodec`` → ``tolokaforge_models.policies.deepseek``

Six of the eight (all except the internal helpers
``JsonRecursiveCoerceResponse`` and ``ItemRecursiveUnwrapResponse``) are
registered with the engine via a new ``tolokaforge.policies`` entry-point
group declared in ``tolokaforge_models/pyproject.toml``.

Engine side (``tolokaforge.core.model_data``) gains
``load_policy_registrations()``: enumerates the entry points, parses each
``"<slot>.<name>"`` key, and returns a ``{slot: {name: class}}`` mapping.
Fail-loud on non-``<slot>.<name>`` keys, non-class values, and duplicate
declarations. ``tolokaforge.core.llm.presets`` calls the loader
immediately after its literal ``_POLICY_REGISTRIES`` initialisation and
merges the returned classes onto each slot; unknown-slot and
class-identity collisions both raise ``RuntimeError``.

Compatibility surface — ``tolokaforge.core.llm.__init__`` gains a
``__getattr__`` deprecation shim mapping the eight moved names to their
new module paths. First access emits ``DeprecationWarning`` naming the
new path; subsequent accesses resolve silently via a module-level
``_WARNED: set[str]`` (Python's default filter would let two callsites
on different lines each re-warn). Shim removed in v0.18.0.

``PER_MODEL_SUBCLASSES`` in ``tests/unit/llm/test_public_api_boundary.py``
updates to the new dotted paths so the four AST-based public-API
guardrails still walk the moved classes' host modules.

Behaviour-lock tests:

* ``tests/unit/llm/test_moved_subclass_shim.py`` (new) — asserts the
  first ``from tolokaforge.core.llm import GeminiRecursiveSchema`` emits
  exactly one ``DeprecationWarning`` naming the new path, a second
  emits zero, a different name emits its own first-time warning, and
  every moved name resolves to the class at its new path.
* ``tests/canonical/test_policy_registry_merge.py`` (new) — locks the
  entry-point declarations, the merged registry contents, the
  unknown-slot fail-loud branch, and the class-identity collision
  branch on the merge helper.
* ``tolokaforge_models/tests/unit/test_policies_reachable.py`` (new) —
  each moved class is importable + instantiable at its new module path.

Per-model tests for the eight subclasses move alongside the classes to
``tolokaforge_models/tests/unit/`` (four test files + one fixture).

Restructured ``tolokaforge_models/`` to a src-layout
(``tolokaforge_models/src/tolokaforge_models/``) so hatchling's editable
install resolves without a ``sources`` prefix rewrite; the wheel-layout
contract stays the same. Root ``pyproject.toml`` adds
``tolokaforge-models>=1.0.0,<2.0.0`` to ``[project].dependencies`` and
declares the workspace source so ``uv sync`` installs both projects
together; the entry-point metadata is written to the models wheel's
``dist-info/entry_points.txt`` at install time.

Wheel-install canonical tests
(``test_adapter_convert_entry_point``, ``test_wheel_install_runner_context``)
now build both wheels into one dist dir and install with ``--find-links``
so the engine wheel's new dep resolves; the runner-import-boundary check
adds ``tolokaforge_models`` to the first-party skip set.
