+++ tolokaforge/core/llm/__init__.py
@@ -3,7 +3,9 @@
 This package is the sole boundary between Tolokaforge and provider-specific
 details (``thinking_blocks``, ``cache_control``, ``reasoning_content``, …).
 Raw provider shapes must never leak above it; callers work with the curated
-public types exported here.
+public types exported here. Per-model policy subclasses ship in
+:mod:`tolokaforge_models.policies` and register with the engine via the
+``tolokaforge.policies`` entry-point group.
 
 See [`docs/LLM_LAYER.md`](../../../docs/LLM_LAYER.md) for module-by-module
 purpose and the plan at
@@ -13,7 +15,7 @@ for design rationale.
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Final
 
 # Safe eager imports — these modules do not depend on tolokaforge.core.models.
 from tolokaforge.core.llm.cache_policy import CachePolicy, NoCache
@@ -32,7 +34,6 @@ from tolokaforge.core.llm.params_policy import GenerationParams, ParamPolicy, Pa
 from tolokaforge.core.llm.prompt_policy import (
     DictMapHints,
     NoPromptEnrichment,
-    RefResolvingDictMapHints,
     SystemPromptPolicy,
 )
 from tolokaforge.core.llm.reasoning import (
@@ -45,25 +46,18 @@ from tolokaforge.core.llm.reasoning_codec import (
     AnthropicReasoningCodec,
     NoReasoningCodec,
     OpenAIReasoningCodec,
-    OpenAISummaryReplayReasoningCodec,
     ReasoningCodec,
 )
 from tolokaforge.core.llm.response_policy import (
     ArrayDictMapResponse,
-    ItemRecursiveUnwrapResponse,
     JsonCoerceResponse,
-    JsonRecursiveCoerceResponse,
-    MinimaxM3TagRecoveryResponse,
     ResponsePolicy,
-    ScalarArrayDictMapResponse,
     StandardResponse,
     UnwrapInputResponse,
     coerce_empty_containers,
     coerce_json_strings,
 )
 from tolokaforge.core.llm.schema_sanitizer import (
-    GeminiRecursiveSchema,
-    GeminiSchema,
     PassthroughSchema,
     SchemaCapability,
     StrictSchema,
@@ -86,7 +80,7 @@ if TYPE_CHECKING:  # pragma: no cover - type-only re-exports
 # and ``build_capabilities`` all import :mod:`tolokaforge.core.models`, which
 # itself imports :mod:`tolokaforge.core.llm.reasoning`. To break the import
 # cycle we expose them via module-level ``__getattr__`` (PEP 562).
-_LAZY: dict[str, tuple[str, str]] = {
+_LAZY: Final[dict[str, tuple[str, str]]] = {
     "ModelCapabilities": ("tolokaforge.core.llm.capabilities", "ModelCapabilities"),
     "LLMClient": ("tolokaforge.core.llm.client", "LLMClient"),
     "UserSimulator": ("tolokaforge.core.llm.client", "UserSimulator"),
@@ -96,7 +90,41 @@ _LAZY: dict[str, tuple[str, str]] = {
 }
 
 
+# ``tolokaforge.core.llm.<name>`` deprecation shim for the eight per-model
+# policy subclasses that moved to :mod:`tolokaforge_models.policies`. Emits
+# ``DeprecationWarning`` on first access per name; subsequent accesses to the
+# same name resolve silently through ``_WARNED``-caching (Python's default
+# filter dedupes on ``(message, category, module, lineno)`` — two callsites
+# on different source lines would each re-warn otherwise). Removed in v0.18.0.
+_MOVED_SUBCLASSES: Final[dict[str, str]] = {
+    "GeminiSchema": "tolokaforge_models.policies.gemini",
+    "GeminiRecursiveSchema": "tolokaforge_models.policies.gemini",
+    "ScalarArrayDictMapResponse": "tolokaforge_models.policies.gemini",
+    "RefResolvingDictMapHints": "tolokaforge_models.policies.inkling",
+    "JsonRecursiveCoerceResponse": "tolokaforge_models.policies.minimax",
+    "ItemRecursiveUnwrapResponse": "tolokaforge_models.policies.minimax",
+    "MinimaxM3TagRecoveryResponse": "tolokaforge_models.policies.minimax",
+    "OpenAISummaryReplayReasoningCodec": "tolokaforge_models.policies.deepseek",
+}
+_WARNED: set[str] = set()
+
+
 def __getattr__(name: str) -> Any:  # pragma: no cover - simple delegator
+    module_path = _MOVED_SUBCLASSES.get(name)
+    if module_path is not None:
+        if name not in _WARNED:
+            import warnings
+
+            warnings.warn(
+                f"tolokaforge.core.llm.{name} moved to {module_path}.{name} — "
+                f"import from the new path. Shim removed in v0.18.0.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            _WARNED.add(name)
+        import importlib
+
+        return getattr(importlib.import_module(module_path), name)
     if name in _LAZY:
         module_name, attr_name = _LAZY[name]
         import importlib
