+++ tolokaforge/core/llm/response_policy.py
@@ -17,13 +17,12 @@ emitted tool-call arguments *after* litellm parsed them:
   :class:`~tolokaforge.core.llm.schema_sanitizer.StrictSchema` array →
   ``Dict[str, T]`` pivot so that tool implementations receive the
   ``Dict[str, T]`` shape they declared.
-- :class:`JsonRecursiveCoerceResponse` — recursive, tags-site-scoped variant
-  of :class:`JsonCoerceResponse` for the MiniMax-M3 ``tags`` corruption.
-- :class:`ItemRecursiveUnwrapResponse` — unwraps the MiniMax-M3 XML
-  repeated-element ``{"item": X}`` artefact into a native list, scoped to the
-  same ``tags`` sites.
-- :class:`MinimaxM3TagRecoveryResponse` — the composite the ``minimax`` preset
-  wires in: ``JsonRecursiveCoerce`` then ``ItemRecursiveUnwrap``.
+
+Per-model response subclasses (``ScalarArrayDictMapResponse``,
+``MinimaxM3TagRecoveryResponse``, ``JsonRecursiveCoerceResponse``,
+``ItemRecursiveUnwrapResponse``) live in
+:mod:`tolokaforge_models.policies` and reach the engine's registry via
+the ``tolokaforge.policies`` entry-point group.
 
 All policies accept an optional ``param_types`` keyword argument: a
 ``Mapping[str, str]`` from root-level parameter name to its JSON-Schema
@@ -47,25 +46,9 @@ __all__ = [
     "UnwrapInputResponse",
     "JsonCoerceResponse",
     "ArrayDictMapResponse",
-    "ScalarArrayDictMapResponse",
-    "JsonRecursiveCoerceResponse",
-    "ItemRecursiveUnwrapResponse",
-    "MinimaxM3TagRecoveryResponse",
     "coerce_empty_containers",
     "coerce_json_strings",
 ]
-
-#: Declared-array ``tags`` sites the MiniMax-M3 recursive recovery policies are
-#: scoped to: the field ``tags`` directly under an ``updates`` parent
-#: (``zendesk_update_item``) or an ``item`` parent (``zendesk_create_item``).
-#: Each entry is a ``(parent_param, field)`` pair. Scoping is mandatory: a
-#: schema-agnostic empty-string → ``[]`` coercion was proven net-harmful — on
-#: MiniMax-M2.7 it produced false-positive scalar corruptions (``''`` on
-#: ``resolution_category__c`` / ``employee_id`` / ``keyword``). The allowlist
-#: ties the recovery to the only paths where ``tags`` is a declared array
-#: inside the schemaless ``additionalProperties: true`` ``updates`` / ``item``
-#: object, so the empty-string guard can never fire on a scalar field.
-ARRAY_SITES: frozenset[tuple[str, str]] = frozenset({("updates", "tags"), ("item", "tags")})
 
 
 def coerce_empty_containers(
@@ -326,293 +309,3 @@ class ArrayDictMapResponse:
                 key = str(value_copy.pop(self.KEY_FIELD))
                 result[param_name] = {key: value_copy}
         return result
-
-
-# ---------------------------------------------------------------------------
-# MiniMax-M3 ``tags`` recovery (recursive, tags-site-scoped)
-# ---------------------------------------------------------------------------
-#
-# MiniMax-M3 routes its tool calls through a provider-side XML → JSON
-# conversion that systematically corrupts the ``tags`` argument: 2505/2505
-# airlines occurrences are malformed. The two dominant shapes are recovered by
-# the two policies below; both are scoped to :data:`ARRAY_SITES` only so the
-# recovery can never touch a scalar field elsewhere in the call. The shipped
-# :class:`JsonCoerceResponse` / :class:`ArrayDictMapResponse` helpers
-# (:func:`coerce_json_strings`, :func:`coerce_empty_containers`) are reused;
-# the new behaviour is *recursion* into the ``updates`` / ``item`` parent plus
-# the site allowlist that bounds it.
-
-
-def _recover_json_string_at_tags_site(value: Any) -> Any:
-    """Recover one ``tags`` value's stringification artefacts.
-
-    Two shapes, both proven on the M3 airlines census:
-
-    * stringified JSON list (``'["a","b"]'``) → the native ``list`` via
-      :func:`coerce_json_strings` (which only promotes ``[`` / ``{`` strings
-      whose ``json.loads`` yields a container — scalar strings are never
-      promoted, by design).
-    * empty string (``''``) → ``[]``. This is the empty-string → ``[]``
-      coercion the spec restricts to declared-array ``tags`` sites: the caller
-      only invokes this helper for paths in :data:`ARRAY_SITES`, so it can
-      never fire on a scalar field (the M2.7 false-positive class).
-
-    Any other value (real list, ``None``, scalar string, number) passes
-    through unchanged.
-
-    Note: the site allowlist is keyed on the field *name* (``tags``), not a
-    verified declared type. It assumes ``updates.tags`` / ``item.tags`` is
-    always a declared array (true for the current mock-tools domains); a
-    future domain declaring a *scalar* field literally named ``tags`` under
-    one of these parents would be mis-coerced.
-    """
-    if value == "" and isinstance(value, str):
-        return []
-    # Reuse the shipped JSON-string decoder by wrapping in a single-key dict;
-    # it never promotes scalar strings and leaves non-string values untouched.
-    return coerce_json_strings({"tags": value})["tags"]
-
-
-# Bound the ``{"item": {"item": ...}}`` unwrap recursion. The observed M3
-# census tops out at one nesting level; greater depth only comes from
-# pathological / adversarial input, so past the cap the value is returned
-# unchanged (refuse to recover rather than raise RecursionError at the
-# tool-call assembly site).
-_MAX_UNWRAP_DEPTH = 64
-
-
-def _unwrap_item_value(value: Any, _depth: int = 0) -> Any:
-    """Recursively unwrap the ``{"item": X}`` XML repeated-element artefact.
-
-    Rule (per spec): a single-key dict ``{"item": X}`` normalises to a list —
-    recurse into ``X`` first (so ``{"item": {"item": "a"}}`` flattens to
-    ``["a"]``), then return ``X`` as-is when it is already a list, else
-    ``[X]``. Multi-key dicts are **left unchanged** (no guessing which key is
-    the real value). Non-dict values pass through untouched.
-    """
-    if not isinstance(value, dict) or set(value.keys()) != {"item"}:
-        return value
-    if _depth >= _MAX_UNWRAP_DEPTH:
-        return value
-    inner = _unwrap_item_value(value["item"], _depth + 1)
-    return inner if isinstance(inner, list) else [inner]
-
-
-class JsonRecursiveCoerceResponse:
-    """Recursive, tags-site-scoped variant of :class:`JsonCoerceResponse`.
-
-    MiniMax-M3's XML → JSON tool-call conversion serialises the ``tags`` array
-    as a JSON-encoded *string* (``'["receipt-issued"]'``, 550/2505 ≈ 22 %) or
-    as an empty string (``''``, 27/2505 ≈ 1 %). Unlike the shipped flat
-    :class:`JsonCoerceResponse`, the corrupt ``tags`` lives **one level deep**
-    inside the schemaless ``additionalProperties: true`` ``updates`` / ``item``
-    object, so recovery has to recurse into that parent.
-
-    Scope (load-bearing): only the ``(parent, field)`` paths in
-    :data:`ARRAY_SITES` are touched — ``updates.tags`` and ``item.tags``. The
-    empty-string → ``[]`` coercion is therefore tied to declared-array sites
-    and can never fire on a scalar field. A schema-agnostic empty-string → ``[]``
-    was proven net-harmful: on MiniMax-M2.7 it corrupted scalar fields such as
-    ``resolution_category__c`` / ``employee_id`` / ``keyword``.
-
-    Never promotes scalar strings (delegates to :func:`coerce_json_strings`),
-    never touches ``None`` / ``null`` (passes straight through).
-    """
-
-    def parse_arguments(
-        self,
-        arguments: dict[str, Any],
-        *,
-        param_types: Mapping[str, str] | None = None,
-    ) -> dict[str, Any]:
-        del param_types  # scoping comes from ARRAY_SITES, not declared types
-        if not isinstance(arguments, dict):
-            return arguments
-        out = dict(arguments)
-        for parent, field in ARRAY_SITES:
-            container = out.get(parent)
-            if not isinstance(container, dict) or field not in container:
-                continue
-            new_container = dict(container)
-            new_container[field] = _recover_json_string_at_tags_site(container[field])
-            out[parent] = new_container
-        return out
-
-
-class ItemRecursiveUnwrapResponse:
-    """Unwrap the MiniMax-M3 ``{"item": X}`` XML repeated-element artefact.
-
-    The provider's XML → JSON conversion renders a repeated XML element
-    (``<tags><item>a</item><item>b</item></tags>``) as a single-key dict keyed
-    on ``item`` rather than a JSON array — 1901/2505 ≈ 76 % of corrupt M3
-    airlines ``tags``, the single largest shape. This policy normalises
-    ``{"item": X}`` to a list (see :func:`_unwrap_item_value`):
-
-    * ``{"item": "receipt-issued"}`` → ``["receipt-issued"]``
-    * ``{"item": ["a", "b"]}`` → ``["a", "b"]`` (already a list — kept flat)
-    * ``{"item": {"item": "a"}}`` → ``["a"]`` (recurses first, then flattens)
-
-    Multi-key dicts (e.g. ``{"item": "a", "refund-requested": ""}``) are left
-    unchanged — there is no safe way to guess which key is the value, so we
-    refuse rather than corrupt (AGENTS.md rule #1).
-
-    Scoped to :data:`ARRAY_SITES` (``updates.tags`` / ``item.tags``) only.
-    """
-
-    def parse_arguments(
-        self,
-        arguments: dict[str, Any],
-        *,
-        param_types: Mapping[str, str] | None = None,
-    ) -> dict[str, Any]:
-        del param_types  # scoping comes from ARRAY_SITES, not declared types
-        if not isinstance(arguments, dict):
-            return arguments
-        out = dict(arguments)
-        for parent, field in ARRAY_SITES:
-            container = out.get(parent)
-            if not isinstance(container, dict) or field not in container:
-                continue
-            new_container = dict(container)
-            new_container[field] = _unwrap_item_value(container[field])
-            out[parent] = new_container
-        return out
-
-
-class MinimaxM3TagRecoveryResponse:
-    """Composite MiniMax-M3 ``tags`` recovery — the ``minimax`` preset's policy.
-
-    A preset has a single ``response_policy`` slot. Following the
-    :class:`ArrayDictMapResponse` precedent (one named policy composing two
-    independent transforms), this composite chains the two M3 ``tags`` recovery
-    stages in order:
-
-    1. :class:`JsonRecursiveCoerceResponse` — stringified-list → ``list`` and
-       ``''`` → ``[]`` at the tags sites. Runs first so a stringified list
-       becomes a real list before unwrapping, and ``''`` becomes ``[]`` rather
-       than being mistaken for anything else.
-    2. :class:`ItemRecursiveUnwrapResponse` — ``{"item": X}`` → list at the
-       tags sites.
-
-    Both stages are scoped to :data:`ARRAY_SITES`; both are no-ops on a valid
-    ``list[str]`` (the order is irrelevant for an already-recovered list, so
-    valid tags pass through unchanged — zero false positives). Gated to the
-    ``minimax`` preset (``minimax-m3*``) only; other models never see it.
-    """
-
-    def __init__(self) -> None:
-        self._coerce = JsonRecursiveCoerceResponse()
-        self._unwrap = ItemRecursiveUnwrapResponse()
-
-    def parse_arguments(
-        self,
-        arguments: dict[str, Any],
-        *,
-        param_types: Mapping[str, str] | None = None,
-    ) -> dict[str, Any]:
-        coerced = self._coerce.parse_arguments(arguments, param_types=param_types)
-        return self._unwrap.parse_arguments(coerced, param_types=param_types)
-
-
-class ScalarArrayDictMapResponse(ArrayDictMapResponse):
-    """:class:`ArrayDictMapResponse` plus a **scalar** dict-map value unwrap.
-
-    Pairs with
-    :class:`~tolokaforge.core.llm.schema_sanitizer.GeminiRecursiveSchema`'s
-    ``carry_scalar_dict_map_value``: a ``Dict[str, int]`` / ``Dict[str, str]``
-    parameter is sent to the model as ``[{key, value}]`` (the scalar carried
-    under the synthetic ``value`` field). The inherited array → dict pivot turns
-    ``[{"key": "SKU-A", "value": 10}]`` into ``{"SKU-A": {"value": 10}}`` — one
-    step short of the ``{"SKU-A": 10}`` the tool's ``Dict[str, int]`` validator
-    wants. This subclass finishes the job by unwrapping every single-field
-    ``{"value": X}`` map entry back to the bare scalar ``X``.
-
-    The same unwrap also recovers the model's *native* wrapper quirk: a model
-    that ignores the array shape and emits ``{"SKU-A": {"value": 10}}`` directly
-    (observed on ``google/gemini-3.5-flash``, 15/15) lands on the identical
-    post-pivot shape, so both the array path and the native-wrapper path
-    converge here.
-
-    Scoped to ``dict_map`` params (via ``param_types``) so a genuine
-    object-valued map that happens to carry a field literally named ``value``
-    is never touched — and even within a dict-map, only a *single-key*
-    ``{"value": …}`` entry is unwrapped, so a multi-field value object is left
-    intact.
-
-    **Nested dict-maps.** The inherited pivot fires only at the root; a dict-map
-    *nested inside an object* param (``order.lines`` where ``order`` is an
-    object and ``lines`` is the ``Dict[str, T]``) is converted to the array
-    shape on the wire like any other dict-map, but the root-only pivot never
-    reaches it — the model then emits ``order.lines`` as the value-less array
-    and the tool rejects it (observed on ``google/gemini-3.5-flash``
-    ``dict_map__nested_in_object``, 0/15). This subclass recurses one level into
-    each ``object`` param and pivots any nested list whose items *all* carry the
-    synthetic ``key`` field back to a dict (and applies the same scalar-``value``
-    unwrap). The recursion is bounded to ``object`` params and keyed on the
-    ``key``-field item signature — the same schema-loss shape the sanitiser
-    produces — so it is a no-op on a genuine ``list[X]`` whose items don't carry
-    ``key``. Its exact reach across domains is data-bound (which nested fields
-    are dict-maps lives in the domain tool schemas, not the probe), so this arm
-    is flagged for a human domain-scope check before merge.
-    """
-
-    def parse_arguments(
-        self,
-        arguments: dict[str, Any],
-        *,
-        param_types: Mapping[str, str] | None = None,
-    ) -> dict[str, Any]:
-        result = super().parse_arguments(arguments, param_types=param_types)
-        if not param_types:
-            return result
-        out = dict(result)
-        for param_name, value in list(out.items()):
-            declared = param_types.get(param_name)
-            if declared == "dict_map" and isinstance(value, dict):
-                out[param_name] = self._unwrap_scalar_values(value)
-            elif declared == "object" and isinstance(value, dict):
-                out[param_name] = self._pivot_nested_dict_maps(value)
-        return out
-
-    @classmethod
-    def _unwrap_scalar_values(cls, mapping: dict[str, Any]) -> dict[str, Any]:
-        """Unwrap single-field ``{"value": X}`` map entries to the bare scalar."""
-        unwrapped: dict[str, Any] = {}
-        for k, v in mapping.items():
-            if isinstance(v, dict) and set(v.keys()) == {StrictSchema.VALUE_FIELD}:
-                unwrapped[k] = v[StrictSchema.VALUE_FIELD]
-            else:
-                unwrapped[k] = v
-        return unwrapped
-
-    @classmethod
-    def _pivot_nested_dict_maps(cls, obj: dict[str, Any]) -> dict[str, Any]:
-        """Recurse an ``object`` param, pivoting nested dict-map-shaped arrays
-        (``[{key, …}]``) back to ``Dict[str, T]`` and scalar-unwrapping the
-        result. Bounded to the ``key``-field item signature so genuine
-        ``list[X]`` values pass through untouched.
-        """
-        out: dict[str, Any] = {}
-        for field, value in obj.items():
-            if (
-                isinstance(value, list)
-                and value
-                and all(isinstance(item, dict) and cls.KEY_FIELD in item for item in value)
-            ):
-                pivoted: dict[str, Any] = {}
-                for item in value:
-                    item_copy = dict(item)
-                    key = str(item_copy.pop(cls.KEY_FIELD))
-                    pivoted[key] = item_copy
-                out[field] = cls._unwrap_scalar_values(pivoted)
-            elif isinstance(value, dict):
-                # A direct dict field one level in is either a native-wrapper
-                # dict-map (``{k: {"value": N}}``) to scalar-unwrap, or a plain
-                # nested object. Recovery is BOUNDED to this one level - the
-                # depth the observe evidence covers (``order.lines``) - so any
-                # deeper subtree passes through untouched. The scalar unwrap is
-                # a no-op on multi-field values, so plain objects are safe.
-                out[field] = cls._unwrap_scalar_values(value)
-            else:
-                out[field] = value
-        return out
