+++ tolokaforge/core/llm/schema_sanitizer.py
@@ -2,7 +2,10 @@
 
 The :class:`ToolSchemaSanitizer` Protocol is the contract every per-model
 sanitizer implements. :class:`SchemaCapability` enumerates the fine-grained
-capabilities the preset system composes.
+capabilities the preset system composes. Concrete per-model sanitisers
+(``GeminiSchema``, ``GeminiRecursiveSchema``) ship in
+:mod:`tolokaforge_models.policies.gemini`; the engine registers them via
+the ``tolokaforge.policies`` entry-point group.
 
 Design principle — **preserve information by default, fail loudly on hazards**.
 The sanitizer must never silently drop schema information the model relies on
@@ -75,7 +78,6 @@ __all__ = [
     "ToolSchemaSanitizer",
     "PassthroughSchema",
     "StrictSchema",
-    "GeminiSchema",
     "SchemaInvariantError",
 ]
 
@@ -830,149 +832,3 @@ class StrictSchema:
             f"construct {match.group()!r} survived sanitisation at offset "
             f"{start}. Context: …{context}…"
         )
-
-
-class GeminiSchema(StrictSchema):
-    """Schema sanitiser tuned for Google Gemini via OpenRouter.
-
-    Gemini's tool spec is a JSON-Schema *subset* — it does not
-    document support for ``$defs`` / ``$ref`` or ``oneOf`` +
-    ``discriminator``. When these constructs appear in the tool
-    schema, Gemini ignores the declared property names inside the
-    unsupported construct and emits arbitrary description-derived
-    English names instead (verified live 2026-05-20: registered
-    ``qty`` → emitted ``quantity``; registered ``subject`` →
-    emitted ``title``).
-
-    This subclass:
-
-    * Inherits ``StrictSchema``'s ``$ref`` inlining, dict-map →
-      array conversion, Decimal collapse, and RE2-incompatible
-      pattern strip.
-    * Sets ``flatten_oneof_discriminator = True`` so Pydantic
-      ``Annotated[..., Field(discriminator='kind')]`` (which emits
-      ``oneOf`` + ``discriminator``) is collapsed into a single
-      object schema unioning every branch's properties.
-      **Bare ``Union[A, B]``** (Pydantic emits ``anyOf`` without
-      ``discriminator``) is left untouched — Gemini handles inline
-      ``anyOf`` branches correctly at the field-name level, and
-      flattening it caused a 40 % → 0 % logistics domain regression.
-    * Sets ``strip_parameters_root_description = False`` so the
-      redundant Pydantic class-docstring at the parameters root
-      (``"Input model for X tool."``) survives. Stripping it
-      correlated with a ~10 pp drop on flat tools
-      (``d365_api_create_case`` / ``custom_listing_id`` regression). The text is
-      semantically redundant but apparently anchors Gemini's
-      optional-field selection.
-    * Sets ``strip_re2_incompatible_patterns = False`` so the
-      Pydantic-emitted RE2-incompatible pattern values (e.g. the
-      Decimal-string idiom ``"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"``)
-      reach Gemini intact. Gemini doesn't enforce RE2; the pattern
-      is just a format hint. Stripping it correlated with a
-      50 % → 10 % travel_marketplace regression for Gemini 3.5
-      Flash, concentrated on the
-      7 Pydantic-Decimal-string patterns the domain happens to
-      carry on its toolset.
-    """
-
-    flatten_oneof_discriminator = True
-    strip_parameters_root_description = False
-    strip_re2_incompatible_patterns = False
-
-
-class GeminiRecursiveSchema(GeminiSchema):
-    """Gemini sanitiser that tolerates **self-referential** (cyclic) ``$ref``.
-
-    ``StrictSchema.inline_refs_in_tool`` fully inlines every ``$ref`` and raises
-    ``"$ref resolution exceeded depth 16"`` the moment a ``$def`` references
-    itself (a Pydantic recursive model such as ``TreeNode.children:
-    list[TreeNode]`` compiles to exactly this cyclic ``$ref``). The raise
-    fires in-engine *before* the request is sent, so the whole trial fails
-    with a sanitiser ``ValueError`` rather than the model ever seeing the
-    tool. Observed on ``google/gemini-3.5-flash`` as ``test_recursive_ref_
-    tool_call`` failing 0/15 on all four shapes (simple / deep_chain /
-    wide_tree / nested_in_object).
-
-    This subclass inlines every *non-cyclic* ``$ref`` exactly as the parent
-    does, but breaks a genuine cycle by substituting a permissive open-object
-    schema (``{"type": "object", "additionalProperties": true}``) at the point
-    of re-entry. The declared schema is thus finite and Gemini receives a valid
-    tool spec; the *receiving* Pydantic validator still accepts an
-    arbitrary-depth nested tree, so a model that emits the full recursion
-    passes. Cycle detection is by def-name on the active resolution stack — a
-    diamond (the same def referenced twice on *different* branches) still
-    inlines on each branch; only a ref that re-enters a def already being
-    resolved on the current path is pruned.
-    """
-
-    #: Pair the recursive-ref tolerance with scalar dict-map value carriage —
-    #: both are Gemini-observed schema-loss surfaces on the same route.
-    carry_scalar_dict_map_value = True
-
-    @classmethod
-    def inline_refs_in_tool(cls, tool: Any) -> Any:
-        """Override of :meth:`StrictSchema.inline_refs_in_tool` — inlines every
-        non-cyclic ``$ref`` exactly as the base does, but substitutes a
-        permissive open-object schema at any point of cyclic re-entry.
-        """
-        if not isinstance(tool, dict):
-            return tool
-        func = tool.get("function")
-        if not isinstance(func, dict):
-            return tool
-        params = func.get("parameters")
-        if not isinstance(params, dict):
-            return tool
-        defs = params.get("$defs")
-        if not isinstance(defs, dict):
-            return tool
-        resolved_params = cls._inline_refs_cycle_tolerant(params, defs, active=())
-        if isinstance(resolved_params, dict):
-            resolved_params.pop("$defs", None)
-        new_func = dict(func)
-        new_func["parameters"] = resolved_params
-        new_tool = dict(tool)
-        new_tool["function"] = new_func
-        return new_tool
-
-    #: Substituted at the point a cyclic ``$ref`` re-enters a def already on
-    #: the active resolution stack. Permissive so the model can still emit the
-    #: recursive subtree; the receiving Pydantic validator enforces the shape.
-    _CYCLE_STUB: dict[str, Any] = {"type": "object", "additionalProperties": True}
-
-    @classmethod
-    def _inline_refs_cycle_tolerant(
-        cls, schema: Any, defs: dict[str, Any], active: tuple[str, ...]
-    ) -> Any:
-        """Inline ``$ref`` like the parent, but prune a ref that re-enters a
-        def already on the ``active`` resolution stack (a genuine cycle)
-        rather than recursing until the depth cap raises.
-        """
-        if isinstance(schema, list):
-            return [cls._inline_refs_cycle_tolerant(item, defs, active) for item in schema]
-        if not isinstance(schema, dict):
-            return schema
-        ref = schema.get("$ref")
-        if isinstance(ref, str) and ref.startswith("#/$defs/"):
-            target_name = ref.removeprefix("#/$defs/")
-            if target_name in active:
-                # Cyclic self-reference — stop inlining, emit a permissive
-                # open object so the wire schema stays finite.
-                return dict(cls._CYCLE_STUB)
-            target = defs.get(target_name)
-            if target is None:
-                raise ValueError(
-                    f"StrictSchema: $ref {ref!r} points to a missing "
-                    f"$defs entry. Available: {sorted(defs.keys())!r}"
-                )
-            resolved = cls._inline_refs_cycle_tolerant(target, defs, active + (target_name,))
-            siblings = {k: v for k, v in schema.items() if k != "$ref"}
-            if not siblings:
-                return resolved
-            if isinstance(resolved, dict):
-                merged = dict(resolved)
-                for k, v in siblings.items():
-                    merged[k] = cls._inline_refs_cycle_tolerant(v, defs, active)
-                return merged
-            return resolved
-        return {k: cls._inline_refs_cycle_tolerant(v, defs, active) for k, v in schema.items()}
