+++ tolokaforge/core/model_data.py
@@ -15,9 +15,13 @@ The module exposes two related surfaces:
   :mod:`tolokaforge.core.engine_run_state`.
 
 The module has **no first-party imports** — it is safe to import from
-runner-subset code. The fingerprint compute path (which needs the
-resolved preset table, pricing dict, and certificate registry) lives in
-the orchestrator-only sibling :mod:`tolokaforge.core.model_data_fingerprint`.
+runner-subset code. :func:`load_policy_registrations` inlines its
+``importlib.metadata`` enumeration (rather than routing through
+:mod:`tolokaforge.core.plugin_registry`) so this module's import graph stays
+light and free of the presets-side circular reach the plugin registry has via
+:mod:`tolokaforge.core.loop`. The fingerprint compute path (which needs the
+resolved preset table, pricing dict, and certificate registry) lives in the
+orchestrator-only sibling :mod:`tolokaforge.core.model_data_fingerprint`.
 
 ``_DATA_ROOT`` is the internal seam pointing at the bundled data
 directory. Tests monkey-patch this constant to redirect the accessors at
@@ -32,6 +36,7 @@ module constants will be sourced from that wheel's ``__init__``.
 
 from __future__ import annotations
 
+import importlib.metadata
 import importlib.resources
 from pathlib import Path
 from typing import Annotated, Any, Final, Literal
@@ -48,7 +53,12 @@ __all__ = [
     "bundled_pricing_path",
     "bundled_providers_path",
     "decode_models_fingerprint",
+    "load_policy_registrations",
 ]
+
+#: Entry-point group each installed models wheel declares its per-model policy
+#: subclasses under. See :func:`load_policy_registrations`.
+POLICIES_GROUP: Final[str] = "tolokaforge.policies"
 
 
 _DATA_ROOT: Final[Path] = Path(str(importlib.resources.files("tolokaforge.core") / "data"))
@@ -171,6 +181,60 @@ class ModelsFingerprint(BaseModel):
         except InvalidSpecifier as exc:
             raise ValueError(f"not a valid PEP 440 specifier: {value!r}") from exc
         return value
+
+
+def load_policy_registrations() -> dict[str, dict[str, type]]:
+    """Discover per-model policy classes shipped by installed models wheels.
+
+    Each entry point in the :data:`POLICIES_GROUP` group declares its slot and
+    policy name via a dotted key ``"<slot>.<name>"`` (for example
+    ``"schema_sanitizer.gemini_recursive"``); the entry point's value resolves
+    to the class object. :mod:`tolokaforge.core.llm.presets` calls this at
+    module import time and merges the returned mapping onto
+    ``_POLICY_REGISTRIES`` after the engine's built-in defaults have been
+    installed.
+
+    Raises :class:`RuntimeError` when an entry-point key is not the expected
+    ``"<slot>.<name>"`` shape, when two distributions register the same
+    ``<slot>.<name>`` pair (a genuine ambiguity that must fail loud with both
+    providers named), or when an entry point resolves to a non-class value.
+    """
+    seen: dict[str, importlib.metadata.EntryPoint] = {}
+    registrations: dict[str, dict[str, type]] = {}
+    for ep in importlib.metadata.entry_points(group=POLICIES_GROUP):
+        existing = seen.get(ep.name)
+        if existing is not None:
+            raise RuntimeError(
+                f"entry point {ep.name!r} in group {POLICIES_GROUP!r} is "
+                f"registered by two distributions: "
+                f"{_distribution_name(existing)!r} and "
+                f"{_distribution_name(ep)!r}"
+            )
+        seen[ep.name] = ep
+        if "." not in ep.name:
+            raise RuntimeError(
+                f"entry point {ep.name!r} in group {POLICIES_GROUP!r} must use "
+                f"the '<slot>.<name>' shape; got a bare token"
+            )
+        slot, _, policy_name = ep.name.partition(".")
+        if not slot or not policy_name:
+            raise RuntimeError(
+                f"entry point {ep.name!r} in group {POLICIES_GROUP!r} must use "
+                f"the '<slot>.<name>' shape; got slot={slot!r} name={policy_name!r}"
+            )
+        cls = ep.load()
+        if not isinstance(cls, type):
+            raise RuntimeError(
+                f"entry point {POLICIES_GROUP}.{ep.name!r} resolved to "
+                f"{type(cls).__name__}, expected a class"
+            )
+        registrations.setdefault(slot, {})[policy_name] = cls
+    return registrations
+
+
+def _distribution_name(ep: importlib.metadata.EntryPoint) -> str:
+    dist = ep.dist
+    return dist.name if dist is not None else "<unknown distribution>"
 
 
 def decode_models_fingerprint(state: dict[str, Any]) -> ModelsFingerprint | None:
