--- tolokaforge_models/policies/__init__.py
@@ -1,10 +1,0 @@
-"""Per-model policy subclasses registered with the engine.
-
-Later stages populate this package with one module per model family;
-each module re-exports its policy classes from ``__all__`` and
-registers them via the ``tolokaforge.policies`` entry-point group.
-"""
-
-from __future__ import annotations
-
-__all__: list[str] = []
