+++ tolokaforge_models/src/tolokaforge_models/data/model_presets.yaml
